requirements.txt
scriptlock
alkivi-logger
flake8
click
python-keystoneclient
python-swiftclient
